requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
streamlit>=1.30.0
plotly>=5.18.0
scikit-learn>=1.3.0
Pillow>=10.0.0
tensorflow>=2.15.0
